backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
python-multipart
openai
numpy
opencv-python
requests
google-generativeai
beautifulsoup4
lxml
Pillow
cloudscraper
